refactor: refactor code to meet expectations
refactor code based on feedback to improve code performance and to conform to global standards
