+++ server/tests/unit/models/document.test.js
@@ -1,6 +1,6 @@
 import chai from 'chai';
-import models from '../../../server/models';
-import mockData from '../mockData/mockData';
+import models from '../../../../server/models';
+import mockData from '../../mockData/mockData';
 
 const expect = chai.expect;
